+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- GeoJSON to JSON and JSON to GeoJSON conversion
+
 ## [0.2.0] - 2022-11-03
 
 ### Changed